+++ assets/scripts/index.js
@@ -23,7 +23,7 @@ require("./twitter-tracking")();
 require("./fetch-packages")();
 require("./tabs")();
 require("./switch-submission")();
-require("./org-in-use-error")();
+require("./validator")();
 
 window.github = require("./github")();
 window.star = require("./star")();
