+++ assets/scripts/validator.js
